Move linecache code to new file, use super()
