+++ bpython/repl.py
@@ -47,8 +47,9 @@ from bpython._py3compat import PythonLexer, py3, prepare_for_exec
 from bpython.clipboard import get_clipboard, CopyFailed
 from bpython.config import getpreferredencoding
 from bpython.formatter import Parenthesis
-from bpython.history import History, filename_for_console_input
+from bpython.history import History
 from bpython.paste import PasteHelper, PastePinnwand, PasteFailed
+from bpython.patch_linecache import filename_for_console_input
 from bpython.translations import _, ngettext
